+++ pyproject.toml
@@ -8,13 +8,15 @@ version = "0.1.0"
 description = "Aggregate search for historical primary source documents"
 readme = "README.md"
 authors = [
-    { name = "Ryan Echols", email = "ryan@shadylakemedia.com" }
+    { name = "Ryan Echols", email = "ryan@shadylakemedia.com" },
 ]
 requires-python = ">=3.13"
 dependencies = [
+    "fastapi~=0.116.1",
     "httpx~=0.28.1",
     "playwright~=1.54.0",
     "pydantic-settings>=2.10.1,<3",
+    "uvicorn[standard]~=0.35.0",
 ]
 
 [dependency-groups]
