+++ src/historical_sources_search/logging/config-dict.json
@@ -1,0 +1,41 @@
+{
+  "version": 1,
+  "formatters": {
+    "default": {
+      "format": "[%(asctime)s] %(levelname)-8s | %(name)s | %(message)s",
+      "datefmt": "%Y-%m-%d %H:%M:%S"
+    }
+  },
+  "filters": {
+    "less_than_warning": {
+      "()": "historical_sources_search.logging.filter_less_than.FilterLessThanLevel",
+      "level": "WARNING"
+    }
+  },
+  "handlers": {
+    "stdout": {
+      "class": "logging.StreamHandler",
+      "stream": "ext://sys.stdout",
+      "formatter": "default",
+      "level": "DEBUG",
+      "filters": ["less_than_warning"]
+    },
+    "stderr": {
+      "class": "logging.StreamHandler",
+      "stream": "ext://sys.stderr",
+      "formatter": "default",
+      "level": "WARNING"
+    }
+  },
+  "loggers": {
+    "__main__": {"level": "DEBUG"},
+    "__mp_main__": {"level": "DEBUG"},
+    "historical_sources_search": {"level": "DEBUG"},
+    "uvicorn": {"level": "INFO"}
+  },
+  "root": {
+    "level": "WARNING",
+    "handlers": ["stderr", "stdout"]
+  },
+  "disable_existing_loggers": false
+}
